backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Database - MongoDB
motor==3.7.0
pymongo==4.10.1

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation & Config
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Document Generation
python-docx==1.1.2
reportlab==4.2.5

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# AI Integration
openai==1.59.3
google-generativeai>=0.8.0
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
